(autohotkey) small fixes; docs

+++ automatic/autohotkey.portable/tools/chocolateyInstall.ps1
@@ -13,14 +13,14 @@ $packageArgs = @{
 Get-ChocolateyUnzip @packageArgs
 Remove-Item $zip_path -ea 0
 
-Write-Host "Removing ANSI version"
+Write-Host "Removing ANSI-32 version"
 Remove-Item "$toolsPath/AutoHotkeyA32.exe" -ea 0
-if (Get-OSArchitectureWidth 64) {
-    Write-Verbose "Removing x32 version"
+if (Get-OSArchitectureWidth 64 -and ($Env:chocolateyForceX86 -ne 'true')) {
+    Write-Verbose "Removing UNICODE-32 version"
     Remove-Item "$toolsPath/AutoHotkeyU32.exe" -ea 0
     Move-Item "$toolsPath/AutoHotkeyU64.exe" "$toolsPath/AutoHotkey.exe" -Force
 } else {
-    Write-Verbose "Removing x64 version"
+    Write-Verbose "Removing UNICODE-64  version"
     Remove-Item "$toolsPath/AutoHotkeyU64.exe" -ea 0
     Move-Item "$toolsPath/AutoHotkeyU32.exe" "$toolsPath/AutoHotkey.exe" -Force
 }
